Move download_file to its own file
